Separate the types.
