+++ neo/SmartContract/NeoService.cs
@@ -4,6 +4,7 @@ using Neo.Network.P2P.Payloads;
 using Neo.Persistence;
 using Neo.SmartContract.Enumerators;
 using Neo.SmartContract.Iterators;
+using Neo.SmartContract.Native.Tokens;
 using Neo.VM;
 using Neo.VM.Types;
 using System;
@@ -14,13 +15,13 @@ using VMArray = Neo.VM.Types.Array;
 
 namespace Neo.SmartContract
 {
-    public partial class NeoService : StandardService
+    public class NeoService : StandardService
     {
         public NeoService(TriggerType trigger, Snapshot snapshot)
             : base(trigger, snapshot)
         {
-            Register("Neo.Native.Tokens.NEO", NeoToken_Main);
-            Register("Neo.Native.Tokens.GAS", GasToken_Main);
+            Register(NeoToken.ServiceName, NeoToken.Main);
+            Register(GasToken.ServiceName, GasToken.Main);
             Register("Neo.Blockchain.GetAccount", Blockchain_GetAccount, 100);
             Register("Neo.Blockchain.GetValidators", Blockchain_GetValidators, 200);
             Register("Neo.Header.GetVersion", Header_GetVersion, 1);
